// src/API.ts
export type S3ImageInput = {
  bucket: string,
  region: string,
  key: string,
};
// ...
export type User = {
  __typename: "User",
  id: string,
  username: string,
  name?: string | null,
  email?: string | null,
  image?: S3Image | null,
  tvShows?: Array< number | null > | null,
  movies?: Array< number | null > | null,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};

export type S3Image = {
  __typename: "S3Image",
  bucket: string,
  region: string,
  key: string,
};

export type UpdateUserInput = {
  id: string,
  username?: string | null,
  name?: string | null,
  email?: string | null,
  image?: S3ImageInput | null,
  tvShows?: Array< number | null > | null,
  movies?: Array< number | null > | null,
};
// ...
export type GetUserQuery = {
  getUser?:  {
    __typename: "User",
    id: string,
    username: string,
    name?: string | null,
    email?: string | null,
    image?:  {
      __typename: "S3Image",
      bucket: string,
      region: string,
      key: string,
    } | null,
    tvShows?: Array< number | null > | null,
    movies?: Array< number | null > | null,
    createdAt: string,
    updatedAt: string,
    owner?: string | null,
  } | null,
};